requirements.txt
flask==0.11.1
ics==0.3.1
requests==2.11.0
python-dotenv==0.5.1
